% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_ang.R
\name{vis_ang}
\alias{vis_ang}
\title{Simple visual angle function}
\usage{
vis_ang(size, scrW, resX, vd)
}
\arguments{
\item{size}{size of required item in pixels}

\item{scrW}{screen width in centimeters}

\item{resX}{screen width in pixels}

\item{vd}{visual distanc from screen}
}
\value{
\preformatted{\item{va}{The visual angle in degrees}
}
}
\description{
Simple visual angle calculator for when the size of the screen is provided in centimeters
}
\examples{


}
